requirements.txt
# GEE通用环境数据提取框架 - Python依赖
# 使用方法: pip install -r requirements.txt

# ========== 核心依赖 ==========

# Google Earth Engine
earthengine-api>=0.1.300

# 配置文件
pyyaml>=6.0

# 数据处理
numpy>=1.21.0
pandas>=1.3.0

# ========== 地理分析 ==========

geemap>=0.1.250
geopandas>=0.12.0
shapely>=2.0.0

# ========== 可视化 ==========

matplotlib>=3.4.0
seaborn>=0.11.0

# ========== Jupyter ==========

jupyter>=1.0.0
ipykernel>=6.0.0

# ========== 进度显示 ==========

tqdm>=4.62.0

# ========== 测试（开发用） ==========

# pytest>=7.2.0
# pytest-cov>=4.0.0
